requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
yarl==1.9.4